Added basic CRUD operations to routes
Added basic CRUD operations to the file carCollectionsRoutes.js to test
out with frontend code. Also added the neccessary code to the
server.js file in order for the routes to work as expected. In addition i
also tried to add som structure to the server.js file. Please Feel free to change
it if you don't like it.